+++ src/main/java/fr/traqueur/currencies/providers/ElementalTokensProvider.java
@@ -4,19 +4,21 @@ import fr.traqueur.currencies.CurrencyProvider;
 import me.elementalgaming.ElementalTokens.TokenAPI;
 import org.bukkit.OfflinePlayer;
 
+import java.math.BigDecimal;
+
 public class ElementalTokensProvider implements CurrencyProvider {
     @Override
-    public void deposit(OfflinePlayer player, double amount) {
-        TokenAPI.addTokens(player.getUniqueId(), (long) amount);
+    public void deposit(OfflinePlayer player, BigDecimal amount) {
+        TokenAPI.addTokens(player.getUniqueId(), amount.longValue());
     }
 
     @Override
-    public void withdraw(OfflinePlayer player, double amount) {
-        TokenAPI.removeTokens(player.getUniqueId(), (long) amount);
+    public void withdraw(OfflinePlayer player, BigDecimal amount) {
+        TokenAPI.removeTokens(player.getUniqueId(), amount.longValue());
     }
 
     @Override
-    public double getBalance(OfflinePlayer player) {
-        return TokenAPI.getTokens(player.getUniqueId());
+    public BigDecimal getBalance(OfflinePlayer player) {
+        return BigDecimal.valueOf(TokenAPI.getTokens(player.getUniqueId()));
     }
 }
